XDrift {xdrift}	R Documentation

XDrift: Spray-drift Modeling Tools for Landscape-Scale Ecotoxicological Risk Assessment

Description

The X3 package currently provides topological functions and spray-drift models.

Topological functions

The bands function can be used to subdivide a plane into parallel bands of equal width. This is useful for applying variability of spray-drift models along field edges. The mindwdist function calculates the minimum distances within a set of 2D-points where some points are defined as sources and the rest are considered possible sinks. Only distances along a wind trajectory are taken into consideration, so that the function returns the minimum distances of sinks downwind of sources. This function can be used to determine source-sink relationships in a spatial context.

Spray-drift models

The rautmann90 is an implementation of the official spray-drift exposure model developed by D. Rautmann [1]. The xspraydrift model is based on the same data as the rautmann90 model but returns values derived from density functions fitted independently for different groups of measurements at discrete distances. This allows incorporating variability of spray-drift exposure into simulations. The model is complemented by the xspraydriftparameters function that lists the parameter sets used by the xspraydrift model.

References

[1] Abdrifteckwerte v. 27.03.2006 edited by Dirk Rautmann JKI.

[Package xdrift version 1.1.0 Index]