bands {XDrift}	R Documentation

Dividing a plane into bands

Description

The bands function is a topological function that subdivides a Euclidean plane into parallel bands of equal width. It returns, for a given set of points, the membership of each point to an individual band.

Usage

bands(x, y, angle, width = 1, angle.def = c("mathematical", "meteorological"))

Arguments

x, y
Numeric vectors of point coordinates. Must have the same length. NAs are not allowed.

angle
A numeric scalar in degrees specifying the orientation of bands. NA is not allowed.

width
A numeric scalar specifying the width of bands in the coordinate system of the point coordinates. Must be positive. NA is not allowed.

angle.def
Specifies the definition of angles used. "mathematical" angles increase counter-clockwise from the direction of the x-axis, "meteorological" angles increase clockwise from the counter-direction of the y-axis.

Details

The bands function works on a set of points on a two-dimensional Euclidean plane. Points are given by their x- and y-coordinates. The plane is subdivided into parallel bands of given width and angle. The angle can either be specified mathematically, i.e., 0 degrees is facing along the x-axis and increasing counter-clockwise, or meteorologically, i.e., 0 degree is facing in opposite direction from the y-axis (wind from the North blowing towards the South) and increasing clockwise. Arbitrary angles can be passed to the function, but they will be internally transformed to mathematical angles in the interval [0 degree; 180 degrees). The intersection of straight lines going through the points with the x- or y-axis is then calculated. Whether the intersection with the x- or y-axis is calculated, is decided on the angle and aims at limiting the inscribed angle to the selected axis to a maximum of 45 degrees. This step prevents precision issues that occur when using trigonometric functions with steeper angles. The intersection is rounded to the specified width and numbered consecutively. The resulting values determine the band, on which each point is located. Equal values thereby identify the same band and different values identify different bands.

Value

The output is a numeric vector of the length of x and y. Equal values indicate that points are located on the same band and differing outputs that they are located on different bands. This makes the function useful for topological grouping of points.

See Also

Other topological functions: mindwdist()

Examples

x <- rep(1:10, 10)
y <- rep(1:10, each = 10)
bands(x, y, 45, width = .7)

[Package XDrift version 1.0.9999 Index]