mindwdist {xdrift}	R Documentation

Calculate the minimum downwind distance between sources and sinks

Description

mindwdist calculates the minimum downwind distance between two types of points: sources and sinks.

Usage

mindwdist(
  x,
  y,
  direction,
  source.indices = 1L,
  max.deviation = 5,
  direction.def = c("meteorological", "mathematical")
)

Arguments

x, y
Numeric vectors of receptor coordinates. Must have the same length. NAs are not allowed.

direction
A numeric scalar specifying the direction of wind in degrees. Must be in the range [0 degree; 360 degree). NA is not allowed.

source.indices
An integer vector specifying the indices of the sources. All values must be in the range [1; length(x)]. NAs are not allowed.

max.deviation
A numeric scalar specifying the maximum deviation from the direction when identifying downwind sinks. Must be in the range [0 degree; 360 degree). NA is not allowed.

direction.def
Specifies the definition of angles. "mathematical" angles are increased counter-clockwise from the direction of the x-axis, "meteorological" angles are increased clockwise from the counter-direction of the y-axis.

Details

The mindwdist function works with a set of 2D-points given by their point coordinates, x and y. Each point is either a source or a possible sink where the source.indices vector specifies the indices of the coordinates that describe sources. Internally, source and sink receptors are represented in matrices, each showing the relation between an individual pair of source and sink. Two relational properties are considered this way: the Euclidean distance between the two points and the angle between the x-axis and the straight line passing through both. The angle is used to identify the points that are located downwind with the wind direction given by the direction argument. Only sinks that deviate not more than the max.deviation in degrees from the direction are considered downwind. The direction can be supplied according to a meteorological definition, i.e., the direction in which the wind blows (0 degrees is South and angles increase clockwise), or in a mathematical way, i.e., 0 degrees is along the x-axis and angles increase counter-clockwise. For each sink, the smallest distance to any source upwind is reported. It is therefore a combination of distance and angle that defines pairs of source-sink-relationships with minimum distance, ignoring sinks that have no source upwind as well as sources that are not nearest to any sink receptor.

Value

A vector of the length of x and y that shows the Euclidean distance of each point to the nearest source in units of the point coordinate system. Source receptors will have a minimum distance of 0. Sinks not located downwind of a source are reported with NA.

See Also

Other topological functions: bands()

Examples

x <- 0:9
y <- rep(0, 10)
mindwdist(x, y, 270, 2:4)

[Package xdrift version 1.1.0 Index]