rautmann90 {xdrift}	R Documentation

Calculating spray-drift exposure according to Rautmann (90th percentile)

Description

The rautmann90 model calculates the spray-drift exposure of sinks downwind of an application source. The function is based on the Rautmann 90th percentile regressions published in Rautmann D, Streloke M, Winkler R. 2001. New basic drift values in the authorisation procedure for plant protection products. In: Forster, R.; Streloke, M.; Workshop on Risk Assessment and Risk Mitigation Measures in the Context of the authorization of Plant Protection Products (WORMM). Germany: Mitt. Biol. Bundesanst. Land-Forstwirtsch. Berlin-Dahlem, Heft 81.

Usage

rautmann90(
  distance,
  crop = c("arable", "vines", "orchards.early", "orchards.late", "hops"),
  target.exposure = NA_real_
)

Arguments

distance
A numeric vector of distances in meters. May contain NAs.

crop
Specifies the crop that is applied. Must be "arable", "vines", "orchards.early", "orchards.late" or "hops".

target.exposure
A numeric scalar specifying the exposure that is reported for distances of 0.

Details

The rautmann90 function works with a set of distances in meters. For each distance, the fraction of spray-drift-deposition relative to the application rate is calculated, taking into account the different formulas for different crops. Negative distances result in an exposure fraction of 0 as they are considered upwind from the exposure source. For distances of exactly 0, cosnidered as representing sources of exposure, the value of the target.exposure scalar is returned, NA by default. Positive distances outside the valid range ([1;250]m for arable and [3;250]m for all other crops) produce warnings.

The rautmann90 model uses composite functions for orchards and hops. The parameters for near and far distances are taken from the Bundesanzeiger and the infliction points, i.e., the distances where to switch from the near-distance parameter set to the far-distance parameter set, were calculated as the root of near and far functions.

Value

A vector of the length of distance that indicates the fraction of exposure individual sinks receive at the given distances. Distances of 0 are reported with the value of target.exposure. Negative distances result in a fraction of 0. NA-distances receive also a fraction of 0 exposure.

See Also

Other spray-drift function: xspraydrift()

Examples

distances <- c(NA, 0:49)
rautmann90(distances)

[Package xdrift version 1.1.0 Index]