xspraydriftparameters {XDrift}	R Documentation

XSprayDrift distribution parameters

Description

Lists the parameters used by the xspraydrift function. For details, seeXDrift – An R Package to Simulate Spatially Explicit Pesticide Spray-drift Exposure of Non-Target-Species Habitats at Landscape Scales, Supplementary Online Material.

Usage

xspraydriftparameters(
  crop = c("orchards.late", "orchards.early", "vines", "arable", "hops"),
  pdf.type = c("gamma", "normal", "lognormal", "beta", "uniform"),
  group = NULL
)

Arguments

crop
Specifies the crop. Must be "orchards.late", "orchards.early", "arable", "hops" or "vines".

pdf.type
Specifies the distribution type. Must be "gamma", "normal", "lognormal", "beta" or "uniform". "uniform" is only supported in combination with "orchards.late".

group
Allows to specify an a priori group for which to return parameters. Possible values are "1-1-1", "1-1-2", "1-1-3", "1-1-4", "1-2-1", "1-2-2", "1-3-1", "2-1-1", "2-1-2", "2-1-4", "2-2-1", "2-2-2", "2-2-3" and "2-3-1", although not all groups are available for all crops. Refer to XSDrift – An R Package to Simulate Spatially Explicit Pesticide Spray-drift Exposure of Non-Target-Species Habitats at Landscape Scales, Supplementary Online Material for details. If group is NULL, the parameters for the lumped distribution for the specified crop are returned.

Details

The xspraydrift model is based on probability distributions derived from field measurements for different distances to the field edge for different crops. It receives the distribution parameters and distance classes from a call of the xspraydriftparameters function for the requested crop and pdf.type. If a group is specified, the parameters for the specified a priori group of crop, pressure class, wind speed class and technology are returned. If no group is specified, the parameters for the lumped distribution of the specified crop are returned.

Value

The output is organized in a nested list. The outer list contains the distribution parameters, i.e., shape and rate for gamma distributions, mean and sd for normal and lognormal distributions, and alpha and beta for beta distributions. The inner list contains predefined distance delineations that specify to which distance to the field edge the parameters apply.

See Also

xspraydrift for applying the parameters

Examples

xspraydriftparameters()

[Package XDrift version 1.0.9999 Index]